Add crowd size relationship with game state.
Now dead pawns are destroyed.
Game can end when no pawns left.
If player wins, the score is number of pawns left.

+++ sources/GUIManager.h
@@ -17,7 +17,7 @@ public:
 
 	void SetUILevel();
 	void SetUIMenu();
-	void SetUIResult(INDEX CurrentLevel, UINT64 Result);
+	void SetUIResult(INDEX CurrentLevel, INT Result);
 
 	std::optional<math::vec4> UpdateUILevel(vec2 cursor, const BOOL InIsReleased);
 	void UpdateUIMenu() {};
